+++ cms-20240330/alibabacloud_cms20240330/models/_create_service_request.py
@@ -2,6 +2,9 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
+from typing import List
+
+from alibabacloud_cms20240330 import models as main_models
 from darabonba.model import DaraModel
 
 class CreateServiceRequest(DaraModel):
@@ -11,9 +14,11 @@ class CreateServiceRequest(DaraModel):
         description: str = None,
         display_name: str = None,
         pid: str = None,
+        resource_group_id: str = None,
         service_name: str = None,
         service_status: str = None,
         service_type: str = None,
+        tags: List[main_models.CreateServiceRequestTags] = None,
     ):
         # Extended attributes.
         self.attributes = attributes
@@ -23,6 +28,7 @@ class CreateServiceRequest(DaraModel):
         self.display_name = display_name
         # Application ID, generally not required to be specified.
         self.pid = pid
+        self.resource_group_id = resource_group_id
         # Service name
         # 
         # This parameter is required.
@@ -33,9 +39,13 @@ class CreateServiceRequest(DaraModel):
         # 
         # This parameter is required.
         self.service_type = service_type
+        self.tags = tags
 
     def validate(self):
-        pass
+        if self.tags:
+            for v1 in self.tags:
+                 if v1:
+                    v1.validate()
 
     def to_map(self):
         result = dict()
@@ -54,6 +64,9 @@ class CreateServiceRequest(DaraModel):
         if self.pid is not None:
             result['pid'] = self.pid
 
+        if self.resource_group_id is not None:
+            result['resourceGroupId'] = self.resource_group_id
+
         if self.service_name is not None:
             result['serviceName'] = self.service_name
 
@@ -62,6 +75,11 @@ class CreateServiceRequest(DaraModel):
 
         if self.service_type is not None:
             result['serviceType'] = self.service_type
+
+        result['tags'] = []
+        if self.tags is not None:
+            for k1 in self.tags:
+                result['tags'].append(k1.to_map() if k1 else None)
 
         return result
 
@@ -79,6 +97,9 @@ class CreateServiceRequest(DaraModel):
         if m.get('pid') is not None:
             self.pid = m.get('pid')
 
+        if m.get('resourceGroupId') is not None:
+            self.resource_group_id = m.get('resourceGroupId')
+
         if m.get('serviceName') is not None:
             self.service_name = m.get('serviceName')
 
@@ -87,6 +108,47 @@ class CreateServiceRequest(DaraModel):
 
         if m.get('serviceType') is not None:
             self.service_type = m.get('serviceType')
+
+        self.tags = []
+        if m.get('tags') is not None:
+            for k1 in m.get('tags'):
+                temp_model = main_models.CreateServiceRequestTags()
+                self.tags.append(temp_model.from_map(k1))
+
+        return self
+
+class CreateServiceRequestTags(DaraModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+    ):
+        self.key = key
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.key is not None:
+            result['key'] = self.key
+
+        if self.value is not None:
+            result['value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('key') is not None:
+            self.key = m.get('key')
+
+        if m.get('value') is not None:
+            self.value = m.get('value')
 
         return self